# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Smart Transport API","description":"Real-time Public Transport Tracking & Density Prediction","version":"1.0.0"},"paths":{"/api/v1/auth/register":{"post":{"tags":["auth"],"summary":"Register","description":"Passenger signup with email and password.","operationId":"register_api_v1_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth"],"summary":"Login","description":"Email/password login. Returns JWT.","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/google":{"post":{"tags":["auth"],"summary":"Google Auth","description":"Google OAuth: verify ID token, create or login user.","operationId":"google_auth_api_v1_auth_google_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GoogleAuthRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["auth"],"summary":"Me","description":"Current user profile (requires JWT).","operationId":"me_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/auth/refresh":{"post":{"tags":["auth"],"summary":"Refresh","description":"Refresh JWT (requires valid token).","operationId":"refresh_api_v1_auth_refresh_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/create":{"post":{"tags":["admin"],"summary":"Create Admin","description":"Admin creates driver or admin. Passengers sign up via /auth/register.","operationId":"create_admin_api_v1_create_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string","default":"Default Name","title":"Name"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminCreateUserRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admin/dashboard/summary":{"get":{"tags":["admin"],"summary":"Dashboard Summary","description":"Counts: active assignments, vehicles, routes, users, raw_telemetry (last 24h).","operationId":"dashboard_summary_api_v1_admin_dashboard_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/admin/dashboard/assignments-over-time":{"get":{"tags":["admin"],"summary":"Assignments Over Time","description":"Assignments per day (last N days).","operationId":"assignments_over_time_api_v1_admin_dashboard_assignments_over_time_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"days","in":"query","required":false,"schema":{"type":"integer","maximum":90,"minimum":1,"default":7,"title":"Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admin/dashboard/occupancy-distribution":{"get":{"tags":["admin"],"summary":"Occupancy Distribution","description":"Occupancy levels (0/1/2) distribution from trip_history.","operationId":"occupancy_distribution_api_v1_admin_dashboard_occupancy_distribution_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/admin/dashboard/eta-accuracy":{"get":{"tags":["admin"],"summary":"Eta Accuracy","description":"Heuristic vs ML vs actual (MAE) from model_performance.","operationId":"eta_accuracy_api_v1_admin_dashboard_eta_accuracy_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/admin/dashboard/route-usage":{"get":{"tags":["admin"],"summary":"Route Usage","description":"Trips per route (last N days).","operationId":"route_usage_api_v1_admin_dashboard_route_usage_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"days","in":"query","required":false,"schema":{"type":"integer","maximum":90,"minimum":1,"default":30,"title":"Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admin/dashboard/telemetry-volume":{"get":{"tags":["admin"],"summary":"Telemetry Volume","description":"Telemetry count per hour (last 24h).","operationId":"telemetry_volume_api_v1_admin_dashboard_telemetry_volume_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/admin/ml/status":{"get":{"tags":["admin"],"summary":"Ml Status","description":"ML model loaded status and version.","operationId":"ml_status_api_v1_admin_ml_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/admin/cleanup":{"post":{"tags":["admin"],"summary":"Trigger Cleanup","description":"Run data retention cleanup (raw_telemetry, trip_history).","operationId":"trigger_cleanup_api_v1_admin_cleanup_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/admin/ml/train":{"post":{"tags":["admin"],"summary":"Trigger Training","description":"Trigger ML model retraining from trip_history.","operationId":"trigger_training_api_v1_admin_ml_train_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/admin/settings":{"get":{"tags":["admin"],"summary":"Get Admin Settings","description":"Get runtime settings (use_ml_for_prod, etc).","operationId":"get_admin_settings_api_v1_admin_settings_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]},"put":{"tags":["admin"],"summary":"Update Settings","description":"Update use_ml_for_prod (admin only).","operationId":"update_settings_api_v1_admin_settings_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SettingsUpdateBody"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/telemetry":{"post":{"tags":["tracking"],"summary":"Receive Telemetry","description":"Receive GPS + density data from SIM7600/ESP32-CAM. Returns immediately; raw save in background.","operationId":"receive_telemetry_api_v1_telemetry_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TelemetryInput"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/assignments/start":{"post":{"tags":["tracking"],"summary":"Start Assignment","description":"Driver check-in: start trip with vehicle and route.","operationId":"start_assignment_api_v1_assignments_start_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignmentStart"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/assignments/end":{"post":{"tags":["tracking"],"summary":"End Assignment","description":"End trip.","operationId":"end_assignment_api_v1_assignments_end_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignmentEnd"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/vehicles":{"post":{"tags":["vehicles"],"summary":"Register Vehicle","operationId":"register_vehicle_api_v1_vehicles_post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehicleCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehicleResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["vehicles"],"summary":"List Vehicles","operationId":"list_vehicles_api_v1_vehicles_get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VehicleResponse"},"title":"Response List Vehicles Api V1 Vehicles Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/vehicles/{vehicle_id}":{"get":{"tags":["vehicles"],"summary":"Get Vehicle","operationId":"get_vehicle_api_v1_vehicles__vehicle_id__get","parameters":[{"name":"vehicle_id","in":"path","required":true,"schema":{"type":"integer","title":"Vehicle Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehicleResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/stops":{"post":{"tags":["routes"],"summary":"Create Stop","operationId":"create_stop_api_v1_stops_post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StopCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StopResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["routes"],"summary":"List Stops","operationId":"list_stops_api_v1_stops_get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StopResponse"},"title":"Response List Stops Api V1 Stops Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/stops/{stop_id}":{"get":{"tags":["routes"],"summary":"Get Stop","operationId":"get_stop_api_v1_stops__stop_id__get","parameters":[{"name":"stop_id","in":"path","required":true,"schema":{"type":"integer","title":"Stop Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StopResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/routes":{"post":{"tags":["routes"],"summary":"Create Route","operationId":"create_route_api_v1_routes_post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["routes"],"summary":"List Routes","operationId":"list_routes_api_v1_routes_get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RouteResponse"},"title":"Response List Routes Api V1 Routes Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/routes/{route_id}":{"get":{"tags":["routes"],"summary":"Get Route","operationId":"get_route_api_v1_routes__route_id__get","parameters":[{"name":"route_id","in":"path","required":true,"schema":{"type":"integer","title":"Route Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteWithStops"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admin/use-ml":{"get":{"tags":["admin"],"summary":"Get Ml Toggle","description":"Get current ETA mode (heuristic vs ML).","operationId":"get_ml_toggle_api_v1_admin_use_ml_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/search/point-to-point":{"post":{"tags":["search"],"summary":"Point To Point Search","description":"Find routes passing through start and end stops.\nReturns routes with pre-calculated bus ETAs from Redis.","operationId":"point_to_point_search_api_v1_search_point_to_point_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PointToPointSearch"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/favorites":{"post":{"tags":["favorites"],"summary":"Add Favorite","operationId":"add_favorite_api_v1_favorites_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FavoriteCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/favorites/{user_id}":{"get":{"tags":["favorites"],"summary":"List Favorites","operationId":"list_favorites_api_v1_favorites__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/ratings":{"post":{"tags":["favorites"],"summary":"Add Rating","operationId":"add_rating_api_v1_ratings_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RatingCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/ratings/{assignment_id}":{"get":{"tags":["favorites"],"summary":"List Ratings","operationId":"list_ratings_api_v1_ratings__assignment_id__get","parameters":[{"name":"assignment_id","in":"path","required":true,"schema":{"type":"integer","title":"Assignment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/notifications/settings":{"post":{"tags":["notifications"],"summary":"Set Notification","operationId":"set_notification_api_v1_notifications_settings_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationSettingCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/notifications/settings/{user_id}":{"get":{"tags":["notifications"],"summary":"Get Notification Settings","operationId":"get_notification_settings_api_v1_notifications_settings__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AdminCreateUserRequest":{"properties":{"username":{"type":"string","maxLength":100,"minLength":3,"title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","maxLength":100,"minLength":8,"title":"Password"},"role":{"type":"string","pattern":"^(driver|admin)$","title":"Role"}},"type":"object","required":["username","email","password","role"],"title":"AdminCreateUserRequest","description":"Admin creates driver or admin."},"AssignmentEnd":{"properties":{"assignment_id":{"type":"integer","title":"Assignment Id"}},"type":"object","required":["assignment_id"],"title":"AssignmentEnd","description":"End an assignment."},"AssignmentStart":{"properties":{"driver_id":{"type":"integer","title":"Driver Id"},"vehicle_id":{"type":"integer","title":"Vehicle Id"},"route_id":{"type":"integer","title":"Route Id"}},"type":"object","required":["driver_id","vehicle_id","route_id"],"title":"AssignmentStart","description":"Start a new assignment (driver check-in)."},"FavoriteCreate":{"properties":{"user_id":{"type":"integer","title":"User Id"},"route_id":{"type":"integer","title":"Route Id"},"nickname":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Nickname"}},"type":"object","required":["user_id","route_id"],"title":"FavoriteCreate"},"GoogleAuthRequest":{"properties":{"id_token":{"type":"string","title":"Id Token"}},"type":"object","required":["id_token"],"title":"GoogleAuthRequest","description":"Google OAuth ID token from frontend."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"LoginRequest","description":"Email/password login."},"NotificationSettingCreate":{"properties":{"user_id":{"type":"integer","title":"User Id"},"route_id":{"type":"integer","title":"Route Id"},"lead_time_minutes":{"type":"integer","title":"Lead Time Minutes","default":10}},"type":"object","required":["user_id","route_id"],"title":"NotificationSettingCreate"},"PointToPointSearch":{"properties":{"start_stop_id":{"type":"integer","title":"Start Stop Id"},"end_stop_id":{"type":"integer","title":"End Stop Id"}},"type":"object","required":["start_stop_id","end_stop_id"],"title":"PointToPointSearch","description":"Search for buses between two stops."},"RatingCreate":{"properties":{"user_id":{"type":"integer","title":"User Id"},"assignment_id":{"type":"integer","title":"Assignment Id"},"score":{"type":"integer","title":"Score"},"comment":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Comment"}},"type":"object","required":["user_id","assignment_id","score"],"title":"RatingCreate"},"RegisterRequest":{"properties":{"username":{"type":"string","maxLength":100,"minLength":3,"title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","maxLength":100,"minLength":8,"title":"Password"}},"type":"object","required":["username","email","password"],"title":"RegisterRequest","description":"Passenger signup (email + password)."},"RouteCreate":{"properties":{"route_number":{"type":"string","title":"Route Number"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"origin":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Origin"},"destination":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Destination"},"stops":{"items":{"$ref":"#/components/schemas/RouteStopSchema"},"type":"array","title":"Stops","default":[]}},"type":"object","required":["route_number"],"title":"RouteCreate"},"RouteResponse":{"properties":{"route_number":{"type":"string","title":"Route Number"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"origin":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Origin"},"destination":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Destination"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["route_number","id"],"title":"RouteResponse"},"RouteStopSchema":{"properties":{"stop_id":{"type":"integer","title":"Stop Id"},"sequence_order":{"type":"integer","title":"Sequence Order"}},"type":"object","required":["stop_id","sequence_order"],"title":"RouteStopSchema"},"RouteWithStops":{"properties":{"route_number":{"type":"string","title":"Route Number"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"origin":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Origin"},"destination":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Destination"},"id":{"type":"integer","title":"Id"},"stops":{"items":{"$ref":"#/components/schemas/StopResponse"},"type":"array","title":"Stops","default":[]}},"type":"object","required":["route_number","id"],"title":"RouteWithStops"},"SettingsUpdateBody":{"properties":{"use_ml_for_prod":{"type":"boolean","title":"Use Ml For Prod","default":false}},"type":"object","title":"SettingsUpdateBody"},"StopCreate":{"properties":{"name":{"type":"string","title":"Name"},"lat":{"type":"number","title":"Lat"},"lon":{"type":"number","title":"Lon"},"base_dwell_time":{"type":"integer","title":"Base Dwell Time","default":30},"is_terminal":{"type":"boolean","title":"Is Terminal","default":false},"peak_multiplier":{"type":"number","title":"Peak Multiplier","default":1.5}},"type":"object","required":["name","lat","lon"],"title":"StopCreate"},"StopResponse":{"properties":{"name":{"type":"string","title":"Name"},"lat":{"type":"number","title":"Lat"},"lon":{"type":"number","title":"Lon"},"base_dwell_time":{"type":"integer","title":"Base Dwell Time","default":30},"is_terminal":{"type":"boolean","title":"Is Terminal","default":false},"peak_multiplier":{"type":"number","title":"Peak Multiplier","default":1.5},"id":{"type":"integer","title":"Id"}},"type":"object","required":["name","lat","lon","id"],"title":"StopResponse"},"TelemetryInput":{"properties":{"device_id":{"type":"string","title":"Device Id"},"lat":{"type":"number","title":"Lat"},"lon":{"type":"number","title":"Lon"},"pixel_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Pixel Count"},"raw_payload":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Raw Payload"}},"type":"object","required":["device_id","lat","lon"],"title":"TelemetryInput","description":"Raw data from SIM7600/ESP32-CAM."},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"TokenResponse"},"UserResponse":{"properties":{"username":{"type":"string","title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"role":{"type":"string","title":"Role","default":"passenger"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["username","email","id","created_at"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VehicleCreate":{"properties":{"plate_number":{"type":"string","title":"Plate Number"},"device_id":{"type":"string","title":"Device Id"},"bus_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bus Type"},"capacity":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Capacity"},"is_active":{"type":"boolean","title":"Is Active","default":true}},"type":"object","required":["plate_number","device_id"],"title":"VehicleCreate"},"VehicleResponse":{"properties":{"plate_number":{"type":"string","title":"Plate Number"},"device_id":{"type":"string","title":"Device Id"},"bus_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bus Type"},"capacity":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Capacity"},"is_active":{"type":"boolean","title":"Is Active","default":true},"id":{"type":"integer","title":"Id"}},"type":"object","required":["plate_number","device_id","id"],"title":"VehicleResponse"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}